LayoutTests/js/dom/dom-attribute-getter-setter.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldThrowErrorName(`console.log(typeof Animation.prototype.timeline)`, `TypeError`);
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
